refactor: move SidebarButtonStates into /types/sidebar
