+++ frontend/src/components/Navbar/Sidebar/Sidebar.tsx
@@ -2,8 +2,8 @@
 import Image from 'next/image'
 import React from 'react'
 import logo from '../../../assets/logo/logo.png'
-import { SidebarElementTypes } from '../../../types/sidebar'
-import { SidebarButton, SidebarButtonState } from './SidebarButton/SidebarButton'
+import { SidebarButtonState, SidebarElementTypes } from '../../../types/sidebar'
+import { SidebarButton } from './SidebarButton/SidebarButton'
 
 export type SidebarProps = {
 	/** Path of the current page. */
