+++ frontend/src/types/sidebar.ts
@@ -1,1 +1,7 @@
 export type SidebarElementTypes = 'testator' | 'marriage' | 'heirs' | 'inheritance' | 'succession' | 'final'
+
+export enum SidebarButtonState {
+	ACTIVE = 'active',
+	DEFAULT = 'default',
+	DISABLED = 'disabled',
+}
